# Optimize CPU GenerateProposals op by lazily generating anchors (3-5x faster)